demo_app/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.4
openpyxl==3.1.2
Werkzeug==3.0.1
requests==2.31.0
python-multipart==0.0.6 